requirements.txt
openai~=0.19.0
PySide2~=5.15.2.1
python-dotenv~=0.21.1
requests~=2.26.0
OpenAIAuth~=0.3.2
urllib3~=1.26.14
typing~=3.7.4.3
revChatGPT~=2.3.5
loguru~=0.6.0
tinydb~=4.7.1
asyncio~=3.4.3
aiohttp~=3.8.4
imgkit~=1.2.3
Markdown~=3.4.1
Pillow~=9.4.0
Pygments~=2.10.0
selenium~=4.8.2
pydantic~=1.10.5
toml~=0.10.2
tiktoken~=0.3.0
httpx~=0.23.3
PyQt5~=5.15.9
style~=1.1.0
wcwidth~=0.2.5
pyperclip~=1.8.2